adding AddingObserver event and RemovingObserver event for Scp173
the adding observer is triggered when a user just saw scp 173. for example, if ev.IsAllowed is set to false and the user is the only one who can saw scp173, the peanut can keep moving (it will be not added in the Observers list
The RemovingObserver is the same as the AddingObserver event but there is no a IsAllowed parameter (it's weird in this situation) and it's trigger when the player no longer saw scp173.

special thanks to Bolton who help me to find stupid IL errors
special thing to yamato: i have compil this xmas-2023 version in release build :trollface: