adding AddingObserver event and RemovingObserver event for Scp173 (again)
the adding observer is triggered when a user just saw scp 173. for example, if ev.IsAllowed is set to false and the user is the only one who can saw scp173, the peanut can keep moving (it will be not added in the Observers list)

an example when the ev.IsAllowed is set to false
![image](https://github.com/Exiled-Team/EXILED/assets/99112969/35454fc1-938a-4c89-bc63-deb0d4dbfbf1)

he RemovingObserver is the same as the AddingObserver event but there is no a IsAllowed parameter (it's weird in this situation) and it's trigger when the player no longer saw scp173.

special thanks to Bolton who help me to find stupid IL errors
special thing to yamato: i have compile this xmas-2023 version in release build :trollface:

edit: ho i forgot to tell: i have build all change and tested the events with a different plugin in my personnal server