Create SECURITY.md
Hi!

You have not got a `SECURITY.md` file. A SECURITY file is used to give guidelines to developers who may find exploits which could be harmful to plugins, servers, or even local machines (Servers or Home Computers).

This file gives a quick outline on what kinds of reports are allowed.

Please feel free to edit this file until you are happy.